// prisma/generated/zod/inputTypeSchemas/BookingItemAddOnUncheckedCreateInputSchema.ts
import { Prisma } from '@prisma/client';

import { z } from 'zod';
import { isValidDecimalInput } from './isValidDecimalInput';
import { DecimalJsLikeSchema } from './DecimalJsLikeSchema';

export const BookingItemAddOnUncheckedCreateInputSchema: z.ZodType<Prisma.BookingItemAddOnUncheckedCreateInput> = z.object({
  id: z.string().cuid().optional(),
  bookingItemId: z.string(),
  addOnId: z.string(),
  addOnName: z.string(),
  unitPrice: z.union([z.number(),z.string(),z.instanceof(Prisma.Decimal),DecimalJsLikeSchema,]).refine((v) => isValidDecimalInput(v), { message: 'Must be a Decimal' }),
  quantity: z.number().int().optional(),
  totalPrice: z.union([z.number(),z.string(),z.instanceof(Prisma.Decimal),DecimalJsLikeSchema,]).refine((v) => isValidDecimalInput(v), { message: 'Must be a Decimal' }),
  createdAt: z.coerce.date().optional(),
  updatedAt: z.coerce.date().optional()
}).strict();

export default BookingItemAddOnUncheckedCreateInputSchema;